Rename coda module to compositional.
